Further cleanup of plugin event system
Fix
