--- packages/react-dom/src/events/ReactBrowserEventEmitter.js
@@ -1,203 +1,0 @@
-/**
- * Copyright (c) Facebook, Inc. and its affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- *
- * @flow
- */
-
-import {registrationNameDependencies} from 'legacy-events/EventPluginRegistry';
-import type {DOMTopLevelEventType} from 'legacy-events/TopLevelEventTypes';
-import {
-  TOP_BLUR,
-  TOP_CANCEL,
-  TOP_CLOSE,
-  TOP_FOCUS,
-  TOP_INVALID,
-  TOP_RESET,
-  TOP_SCROLL,
-  TOP_SUBMIT,
-  getRawEventName,
-  mediaEventTypes,
-} from './DOMTopLevelEventTypes';
-import {
-  setEnabled,
-  isEnabled,
-  trapBubbledEvent,
-  trapCapturedEvent,
-} from './ReactDOMEventListener';
-import isEventSupported from './isEventSupported';
-
-/**
- * Summary of `ReactBrowserEventEmitter` event handling:
- *
- *  - Top-level delegation is used to trap most native browser events. This
- *    may only occur in the main thread and is the responsibility of
- *    ReactDOMEventListener, which is injected and can therefore support
- *    pluggable event sources. This is the only work that occurs in the main
- *    thread.
- *
- *  - We normalize and de-duplicate events to account for browser quirks. This
- *    may be done in the worker thread.
- *
- *  - Forward these native events (with the associated top-level type used to
- *    trap it) to `EventPluginRegistry`, which in turn will ask plugins if they want
- *    to extract any synthetic events.
- *
- *  - The `EventPluginRegistry` will then process each event by annotating them with
- *    "dispatches", a sequence of listeners and IDs that care about that event.
- *
- *  - The `EventPluginRegistry` then dispatches the events.
- *
- * Overview of React and the event system:
- *
- * +------------+    .
- * |    DOM     |    .
- * +------------+    .
- *       |           .
- *       v           .
- * +------------+    .
- * | ReactEvent |    .
- * |  Listener  |    .
- * +------------+    .                         +-----------+
- *       |           .               +--------+|SimpleEvent|
- *       |           .               |         |Plugin     |
- * +-----|------+    .               v         +-----------+
- * |     |      |    .    +--------------+                    +------------+
- * |     +-----------.--->|PluginRegistry|                    |    Event   |
- * |            |    .    |              |     +-----------+  | Propagators|
- * | ReactEvent |    .    |              |     |TapEvent   |  |------------|
- * |  Emitter   |    .    |              |<---+|Plugin     |  |other plugin|
- * |            |    .    |              |     +-----------+  |  utilities |
- * |     +-----------.--->|              |                    +------------+
- * |     |      |    .    +--------------+
- * +-----|------+    .                ^        +-----------+
- *       |           .                |        |Enter/Leave|
- *       +           .                +-------+|Plugin     |
- * +-------------+   .                         +-----------+
- * | application |   .
- * |-------------|   .
- * |             |   .
- * |             |   .
- * +-------------+   .
- *                   .
- *    React Core     .  General Purpose Event Plugin System
- */
-
-const PossiblyWeakMap = typeof WeakMap === 'function' ? WeakMap : Map;
-// prettier-ignore
-const elementListenerMap:
-  // $FlowFixMe Work around Flow bug
-  | WeakMap
-  | Map<
-      Document | Element | Node,
-      Map<DOMTopLevelEventType | string, null | (any => void)>,
-    > = new PossiblyWeakMap();
-
-export function getListenerMapForElement(
-  element: Document | Element | Node,
-): Map<DOMTopLevelEventType | string, null | (any => void)> {
-  let listenerMap = elementListenerMap.get(element);
-  if (listenerMap === undefined) {
-    listenerMap = new Map();
-    elementListenerMap.set(element, listenerMap);
-  }
-  return listenerMap;
-}
-
-/**
- * We listen for bubbled touch events on the document object.
- *
- * Firefox v8.01 (and possibly others) exhibited strange behavior when
- * mounting `onmousemove` events at some node that was not the document
- * element. The symptoms were that if your mouse is not moving over something
- * contained within that mount point (for example on the background) the
- * top-level listeners for `onmousemove` won't be called. However, if you
- * register the `mousemove` on the document object, then it will of course
- * catch all `mousemove`s. This along with iOS quirks, justifies restricting
- * top-level listeners to the document object only, at least for these
- * movement types of events and possibly all events.
- *
- * @see http://www.quirksmode.org/blog/archives/2010/09/click_event_del.html
- *
- * Also, `keyup`/`keypress`/`keydown` do not bubble to the window on IE, but
- * they bubble to document.
- *
- * @param {string} registrationName Name of listener (e.g. `onClick`).
- * @param {object} mountAt Container where to mount the listener
- */
-export function listenTo(
-  registrationName: string,
-  mountAt: Document | Element | Node,
-): void {
-  const listeningSet = getListenerMapForElement(mountAt);
-  const dependencies = registrationNameDependencies[registrationName];
-
-  for (let i = 0; i < dependencies.length; i++) {
-    const dependency = dependencies[i];
-    listenToTopLevel(dependency, mountAt, listeningSet);
-  }
-}
-
-export function listenToTopLevel(
-  topLevelType: DOMTopLevelEventType,
-  mountAt: Document | Element | Node,
-  listenerMap: Map<DOMTopLevelEventType | string, null | (any => void)>,
-): void {
-  if (!listenerMap.has(topLevelType)) {
-    switch (topLevelType) {
-      case TOP_SCROLL:
-        trapCapturedEvent(TOP_SCROLL, mountAt);
-        break;
-      case TOP_FOCUS:
-      case TOP_BLUR:
-        trapCapturedEvent(TOP_FOCUS, mountAt);
-        trapCapturedEvent(TOP_BLUR, mountAt);
-        // We set the flag for a single dependency later in this function,
-        // but this ensures we mark both as attached rather than just one.
-        listenerMap.set(TOP_BLUR, null);
-        listenerMap.set(TOP_FOCUS, null);
-        break;
-      case TOP_CANCEL:
-      case TOP_CLOSE:
-        if (isEventSupported(getRawEventName(topLevelType))) {
-          trapCapturedEvent(topLevelType, mountAt);
-        }
-        break;
-      case TOP_INVALID:
-      case TOP_SUBMIT:
-      case TOP_RESET:
-        // We listen to them on the target DOM elements.
-        // Some of them bubble so we don't want them to fire twice.
-        break;
-      default:
-        // By default, listen on the top level to all non-media events.
-        // Media events don't bubble so adding the listener wouldn't do anything.
-        const isMediaEvent = mediaEventTypes.indexOf(topLevelType) !== -1;
-        if (!isMediaEvent) {
-          trapBubbledEvent(topLevelType, mountAt);
-        }
-        break;
-    }
-    listenerMap.set(topLevelType, null);
-  }
-}
-
-export function isListeningToAllDependencies(
-  registrationName: string,
-  mountAt: Document | Element,
-): boolean {
-  const listenerMap = getListenerMapForElement(mountAt);
-  const dependencies = registrationNameDependencies[registrationName];
-
-  for (let i = 0; i < dependencies.length; i++) {
-    const dependency = dependencies[i];
-    if (!listenerMap.has(dependency)) {
-      return false;
-    }
-  }
-  return true;
-}
-
-export {setEnabled, isEnabled, trapBubbledEvent, trapCapturedEvent};
